Updated flow of ticket channel to made dynamic so we can add dynamically ticket too and refactor and made structured services
