+++ src/config/products.js
@@ -1,0 +1,21 @@
+export const ALLOWED_PRODUCTS = [
+  'earthrover',
+  'earthrover_school',
+  'ufb',
+  'sam',
+  'robotsfun',
+  'et_fugi'
+];
+
+export const PRODUCT_DISPLAY_NAMES = {
+  earthrover: 'Earthrover',
+  earthrover_school: 'Earthrover School',
+  ufb: 'UFB',
+  sam: 'SAM',
+  robotsfun: 'Robots.Fun',
+  et_fugi: 'ET Fugi'
+};
+
+export default { ALLOWED_PRODUCTS, PRODUCT_DISPLAY_NAMES };
+
+
