+++ src/services/DynamicPublicChannelService.js
@@ -1,12 +1,16 @@
 import redis from './redisClient.js';
+import EventEmitter from 'events';
 
 /**
  * Dynamic Public Channel Service
  * Manages public channels stored in Redis for real-time updates without restart
  */
-class DynamicPublicChannelService {
+class DynamicPublicChannelService extends EventEmitter {
   constructor() {
+    super();
     this.REDIS_KEY_PREFIX = 'public_channels:';
+    this.cache = new Map(); // guildId -> Set(channelIds)
+    this._interval = null;
   }
 
   /**
@@ -35,6 +39,11 @@ class DynamicPublicChannelService {
       console.log('[Redis] Saving public channel payload:', channelData);
       await redis.hSet(key, channelId, JSON.stringify(channelData));
       console.log(`✅ Added dynamic public channel ${channelId} for guild ${guildId}`);
+      this.emit('channel:add', { guildId, channelId });
+      // Push-update cache
+      const set = this.cache.get(guildId) || new Set();
+      set.add(channelId);
+      this.cache.set(guildId, set);
       return true;
     } catch (error) {
       console.error(`❌ Error adding public channel ${channelId} for guild ${guildId}:`, error);
@@ -50,6 +59,9 @@ class DynamicPublicChannelService {
       const key = this._getRedisKey(guildId);
       const result = await redis.hDel(key, channelId);
       console.log(`✅ Removed dynamic public channel ${channelId} for guild ${guildId}`);
+      this.emit('channel:remove', { guildId, channelId });
+      const set = this.cache.get(guildId);
+      if (set) set.delete(channelId);
       return result > 0;
     } catch (error) {
       console.error(`❌ Error removing public channel ${channelId} for guild ${guildId}:`, error);
@@ -62,6 +74,9 @@ class DynamicPublicChannelService {
    */
   async getDynamicPublicChannels(guildId) {
     try {
+      // Serve from cache first
+      const set = this.cache.get(guildId);
+      if (set && set.size) return Array.from(set);
       const key = this._getRedisKey(guildId);
       const channels = await redis.hGetAll(key);
       
@@ -77,6 +92,8 @@ class DynamicPublicChannelService {
         }
       }
 
+      // Prime cache
+      this.cache.set(guildId, new Set(activeChannels));
       return activeChannels;
     } catch (error) {
       console.error(`❌ Error getting dynamic public channels for guild ${guildId}:`, error);
@@ -104,6 +121,8 @@ class DynamicPublicChannelService {
    */
   async isPublicChannel(guildId, channelId) {
     try {
+      const set = this.cache.get(guildId);
+      if (set) return set.has(channelId);
       const dynamicChannels = await this.getAllPublicChannels(guildId);
       return dynamicChannels.includes(channelId);
     } catch (error) {
@@ -157,6 +176,44 @@ class DynamicPublicChannelService {
     } catch (error) {
       console.error(`❌ Error getting Google Docs links for channel ${channelId}:`, error);
       return [];
+    }
+  }
+
+  // Cache refresher using SCAN to avoid blocking
+  async refreshCache() {
+    try {
+      const prefix = this.REDIS_KEY_PREFIX;
+      let cursor = '0';
+      const guildIds = new Set();
+      do {
+        const res = await redis.scan(cursor, { MATCH: `${prefix}*`, COUNT: 100 });
+        cursor = res.cursor;
+        for (const key of res.keys) {
+          const guildId = key.replace(prefix, '');
+          guildIds.add(guildId);
+        }
+      } while (cursor !== '0');
+
+      for (const guildId of guildIds) {
+        const ids = await this.getDynamicPublicChannels(guildId);
+        this.cache.set(guildId, new Set(ids));
+      }
+    } catch (error) {
+      console.error('❌ Error refreshing public channel cache:', error);
+    }
+  }
+
+  startCacheRefresher(intervalMs = 10000) {
+    if (this._interval) return;
+    this.refreshCache();
+    this._interval = setInterval(() => this.refreshCache(), intervalMs);
+    console.log(`🗂️ DynamicPublicChannelService cache refresher started (every ${intervalMs}ms)`);
+  }
+
+  stopCacheRefresher() {
+    if (this._interval) {
+      clearInterval(this._interval);
+      this._interval = null;
     }
   }
 }
