+++ src/services/dynamic/DynamicPublicChannelService.js
@@ -1,0 +1,3 @@
+export { default } from '../DynamicPublicChannelService.js';
+
+
