+++ src/services/dynamic/DynamicTicketChannelService.js
@@ -1,0 +1,3 @@
+export { default } from '../DynamicTicketChannelService.js';
+
+
