+++ src/utils/TicketChannelUtil.js
@@ -1,10 +1,13 @@
-import { getServerConfig } from '../config/serverConfigs.js';
+import dynamicTicketChannelService from '../services/dynamic/DynamicTicketChannelService.js';
 
 class TicketChannelUtil {
   static isTicketChannel(channel) {
     try {
-      const serverConfig = getServerConfig(channel.guild?.id);
-      return channel.isThread && channel.isThread() && serverConfig && channel.parentId === serverConfig.ticketChannelId;
+      if (!channel.isThread || !channel.isThread()) return false;
+      const guildId = channel.guild?.id;
+      const parentId = channel.parentId;
+      // Dynamic only
+      return dynamicTicketChannelService.isTicketChannelCached(guildId, parentId);
     } catch {
       return false;
     }
